Module Reparse

Overview

Parser provides functions and types to construct robust, performant and reusable parsers. At the core is a type Reparse.t which represents a constructed parser definition. A parser Reparse.t is defined by composing together one or more parsers or Reparse.ts via usage of parser operators. An instance of Reparse.t represents an un-evaluated parser. Use Reparse.parse function to evaluate it. Reparse.input represents a generalization of data input to Reparse.parse. Implement the interface to create new input types. Parser operators - or functions - are broadly organized into following categories:

Monadic parsers

Char/String parsers

Alternate parsers

Boolean

Repetition

Others

module type PARSER = sig ... end

module type INPUT = sig ... end

module Make (Input : INPUT) : PARSER with type 'a promise = 'a Input.promise with type input = Input.t

A functor to create parsers based on the given Input module.

module String : PARSER with type 'a promise = 'a with type input = string

A parser when the input is a string.